--- src/main/java/pl/edu/pk/shop/address/.gitignore
@@ -1,2 +1,0 @@
-/AddressData.class
-/Address.class
