--- src/main/java/pl/edu/pk/shop/address/Address.java
@@ -1,25 +1,0 @@
-package pl.edu.pk.shop.address;
-
-/** Address - represents an instance of address, 
- * specified for User class
- * @author Christopher Abram
- **/
-public class Address {
-	// vars {
-	
-		public AddressData data;
-	
-	// } methods {
-		// public {
-				
-			public Address(int ID){
-				data = new AddressData(ID);
-			}
-	
-		// } protected {
-				
-		// } private {
-				
-		// }
-	// }
-}
