--- src/main/java/pl/edu/pk/shop/function/.gitignore
@@ -1,2 +1,0 @@
-/Function.class
-/FunctionData.class
