--- src/main/java/pl/edu/pk/shop/function/Function.java
@@ -1,25 +1,0 @@
-package pl.edu.pk.shop.function;
-
-/** Function - represents an instance of function, 
- * specified for User class
- * @author Christopher Abram
- **/
-public class Function {
-	// vars {
-		
-		public FunctionData data;
-		
-	// } methods {
-		// public {
-			
-			public Function(int ID){
-				data = new FunctionData(ID);
-			}// end Function
-			
-		// } protected {
-			
-		// } private {
-			
-		// }
-	// }
-}
