--- src/main/java/pl/edu/pk/shop/item/.gitignore
@@ -1,2 +1,0 @@
-/Item.class
-/ItemData.class
