--- src/main/java/pl/edu/pk/shop/item/Item.java
@@ -1,34 +1,0 @@
-package pl.edu.pk.shop.item;
-
-
-
-public class Item {
-	// vars {
-	
-		// Item data:
-		public ItemData data;
-		
-	// } methods {
-		// public {
-			
-			/**
-			 **/
-			public Item(int ID){
-				data = new ItemData(ID);
-			}// end Item
-			
-			/**
-			 **/
-			
-			
-		// } protected {
-			
-			
-			
-		// } private {
-			
-			
-			
-		// }
-	// }
-}
