--- src/main/java/pl/edu/pk/shop/tabledata/.gitignore
@@ -1,1 +1,0 @@
-/TableData.class
