--- src/main/java/pl/edu/pk/shop/user/.gitignore
@@ -1,5 +1,0 @@
-/Admin.class
-/Client.class
-/Guest.class
-/User.class
-/UserData.class
