--- src/main/java/pl/edu/pk/shop/user/Admin.java
@@ -1,24 +1,0 @@
-package pl.edu.pk.shop.user;
-
-/**
- * @author Christopher Abram
- */
-public class Admin extends User {
-	// vars {
-			
-	// } methods {
-			
-		// public {
-				
-			public Admin(int ID){
-				super(ID);
-			}// end Admin
-	
-		// } protected {
-				
-		// } private {
-				
-		// }
-			
-	// }
-}
