--- src/main/java/pl/edu/pk/shop/user/Client.java
@@ -1,24 +1,0 @@
-package pl.edu.pk.shop.user;
-
-/**
- * @author Christopher Abram
- */
-public class Client extends User {
-	// vars {
-			
-	// } methods {
-			
-		// public {
-				
-			public Client(int ID){
-				super(ID);
-			}// end Client
-	
-		// } protected {
-				
-		// } private {
-				
-		// }
-			
-	// }
-}
