--- src/main/java/pl/edu/pk/shop/user/Guest.java
@@ -1,24 +1,0 @@
-package pl.edu.pk.shop.user;
-
-/**
- * @author Christopher Abram
- */
-public class Guest extends User {
-	// vars {
-			
-	// } methods {
-			
-		// public {
-				
-			public Guest(int ID){
-				super(ID);
-			}// end Guest
-	
-		// } protected {
-				
-		// } private {
-				
-		// }
-			
-	// }
-}
